Don't make `sandbox_malloc` into a coroutine
Okay, the coroutine implementation of `sandbox_malloc` is clearly
broken. It would be working if Asyncify instrumented the `memory.grow`
WebAssembly instruction, but it doesn't instrument it.

This commit reverts commit 42c4ff9497519660a2e4fb051a18fabbaa9799b8 and
also increases the default VM memory allocation from 64 MiB to 96 MiB to
account for the lack of ability to increase the memory allocation at run
time. I'll find some new way to implement increasing the memory
allocation later.

+++ src/core.cpp
@@ -257,7 +257,7 @@ static bool init_sandbox() {
             return std::calloc(1, sizeof(long));
         },
         .fread = [](void *buf, int count, void *handle) {
-            assert(*(long *)handle + count < mkxp_gmgsx_sf2_len);
+            assert((size_t)(*(long *)handle + count) < mkxp_gmgsx_sf2_len);
             std::memcpy(buf, mkxp_gmgsx_sf2 + *(long *)handle, count);
             *(long *)handle += count;
             return (int)FLUID_OK;
